Align LLM provider tests with fixtures
