+++ apps/frontman_server/lib/frontman_server.ex
@@ -39,7 +39,7 @@ defmodule FrontmanServer do
   ]
 
   @exports (case Mix.env() do
-              :test -> @base_exports ++ [DataCase, Testing.LLMProviderHelpers]
+              :test -> @base_exports ++ [DataCase, ExecutionCase, Test.Fixtures.LLMProvider]
               _ -> @base_exports
             end)
