+++ apps/frontman_server/test/frontman_server/observability/otel_handler_test.exs
@@ -15,12 +15,11 @@ defmodule FrontmanServer.Observability.OtelHandlerTest do
           └── tool (TOOL)
   ```
   """
-  use SwarmAi.Testing, async: false
+  use FrontmanServer.ExecutionCase
 
   import FrontmanServer.Test.Fixtures.Accounts
   import FrontmanServer.InteractionCase.Helpers
   import FrontmanServer.Test.Fixtures.Tasks
-  import FrontmanServer.Testing.LLMProviderHelpers
 
   alias Ecto.Adapters.SQL.Sandbox
   alias FrontmanServer.Tasks
