+++ apps/frontman_server/test/frontman_server/tasks/execution/mcp_tool_broadcast_test.exs
@@ -6,13 +6,12 @@ defmodule FrontmanServer.Tasks.Execution.MCPToolBroadcastTest do
   from SwarmCase, catching issues like duplicate tool call broadcasts.
   """
 
-  use SwarmAi.Testing, async: false
+  use FrontmanServer.ExecutionCase
 
   import FrontmanServer.InteractionCase.Helpers
 
   import FrontmanServer.Test.Fixtures.Accounts
   import FrontmanServer.Test.Fixtures.Tasks
-  import FrontmanServer.Testing.LLMProviderHelpers
 
   alias Ecto.Adapters.SQL.Sandbox
   alias FrontmanServer.Tasks
