+++ apps/frontman_server/test/support/channel_case.ex
@@ -20,7 +20,7 @@ defmodule FrontmanServerWeb.ChannelCase do
   alias Ecto.Adapters.SQL.Sandbox
   alias FrontmanServer.Accounts
   alias FrontmanServer.Accounts.Scope
-  alias FrontmanServer.Testing.LLMProviderHelpers
+  alias FrontmanServer.Test.Fixtures.LLMProvider
 
   using do
     quote do
@@ -210,7 +210,7 @@ defmodule FrontmanServerWeb.ChannelCase do
       raise "Cannot combine shared_sandbox: true with async: true - shared sandbox requires synchronous execution"
     end
 
-    LLMProviderHelpers.stub_llm_response("Test response")
+    LLMProvider.stub_llm_response("Test response")
 
     shared = tags[:shared_sandbox] || not tags[:async]
     pid = Sandbox.start_owner!(FrontmanServer.Repo, shared: shared)
